Refactor for dashboard prompting
Refactor the code so that the core functionality can be used from the PZP core (as well as via a web-socket request).

Jira issue: wp-956
